# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/accounts": {
            "post": {
                "description": "Creates an account for a document number.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Create account",
                "parameters": [
                    {
                        "description": "Document number",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateAccountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/accounts/{accountId}": {
            "get": {
                "description": "Returns account id, document number, balance, and active installment plans (incomplete).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Get account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetAccountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/accounts/{accountId}/installments/{planId}/pay": {
            "post": {
                "description": "Records a credit voucher for this EMI and advances the plan (does not change the original purchase debit).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Pay one installment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Installment plan ID",
                        "name": "planId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.PayInstallmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions": {
            "post": {
                "description": "Types 1–3 debit (stored negative), type 4 credit (stored positive). For type 2, optional tenure (\u003e1) creates an installment plan; omit tenure for a lump debit only. EMI repayment: POST .../installments/.../pay (credit voucher). Fails if balance would go negative on debits.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Create transaction",
                "parameters": [
                    {
                        "description": "Transaction payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateTransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.CreateAccountRequest": {
            "type": "object",
            "properties": {
                "document_number": {
                    "type": "string",
                    "example": "12345678900"
                }
            }
        },
        "handlers.CreateAccountResponse": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer",
                    "example": 1
                },
                "document_number": {
                    "type": "string",
                    "example": "12345678900"
                }
            }
        },
        "handlers.CreateTransactionRequest": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer",
                    "example": 1
                },
                "amount": {
                    "type": "number",
                    "example": 123.45
                },
                "operation_type_id": {
                    "type": "integer",
                    "example": 1
                },
                "tenure": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "handlers.CreateTransactionResponse": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "amount": {
                    "type": "number"
                },
                "event_date": {
                    "type": "string"
                },
                "installment_plan": {
                    "$ref": "#/definitions/handlers.InstallmentPlanItem"
                },
                "operation_type_id": {
                    "type": "integer"
                },
                "transaction_id": {
                    "type": "integer"
                }
            }
        },
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "record not found"
                }
            }
        },
        "handlers.GetAccountResponse": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer",
                    "example": 1
                },
                "active_installment_plans": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.InstallmentPlanItem"
                    }
                },
                "balance": {
                    "type": "number",
                    "example": 0
                },
                "document_number": {
                    "type": "string",
                    "example": "12345678900"
                }
            }
        },
        "handlers.InstallmentPlanItem": {
            "type": "object",
            "properties": {
                "next_due_date": {
                    "type": "string"
                },
                "paid_emis": {
                    "type": "integer"
                },
                "plan_id": {
                    "type": "integer"
                },
                "remaining_emis": {
                    "type": "integer"
                },
                "tenure": {
                    "type": "integer"
                },
                "total_amount": {
                    "type": "number"
                }
            }
        },
        "handlers.PayInstallmentResponse": {
            "type": "object",
            "properties": {
                "paid_emis": {
                    "type": "integer"
                },
                "remaining_emis": {
                    "type": "integer"
                }
            }
        }
    }
}